% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfunctions.R
\name{all_geweke_tests}
\alias{all_geweke_tests}
\title{The user calls it like:
test_fns <- make_test_functions(
mu_sq    = function(theta, y) theta\link{"mu"}^2,
product  = function(theta, y) theta\link{"mu"} * theta\link{"sigmasq"},
log_like = function(theta, y) sum(dnorm(y, theta\link{"mu"},
sqrt(theta\link{"sigmasq"}), log = TRUE))
)}
\usage{
all_geweke_tests(direct_draws, gibbs_draws, test_functions)
}
\description{
The user calls it like:
test_fns <- make_test_functions(
mu_sq    = function(theta, y) theta\link{"mu"}^2,
product  = function(theta, y) theta\link{"mu"} * theta\link{"sigmasq"},
log_like = function(theta, y) sum(dnorm(y, theta\link{"mu"},
sqrt(theta\link{"sigmasq"}), log = TRUE))
)
}
