% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geweke.R
\name{apply_test_functions}
\alias{apply_test_functions}
\title{apply_test_functions}
\usage{
apply_test_functions(direct_draws, gibbs_draws, test_functions)
}
\arguments{
\item{direct_draws}{Your matrix of direct draws - e.g., test_matrix$direct when test_matrix is your apply_test_functions output.}

\item{gibbs_draws}{Your matrix of gibbs draws - e.g. test_matrix$gibbs.}

\item{test_functions}{Your test functions. To access parameters, access as theta[\link{"name"}] or theta[\link{index}] with a double bracket not single bracket, and unlist(theta\link{1:p}) to extract sub-vector parameters.}
}
\value{
Returns a list of test matrices - one $direct, one $gibbs - with one column for each test function's values.
}
\description{
Takes in the direct and Gibbs draws and applies the specified test functions to output test values.
}
