% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geweke_list.R, R/geweke_matrix.R
\name{as_theta}
\alias{as_theta}
\title{the following two functions create the successive-conditional and
marginal-conditional sampling needed for comparison per John Geweke's 2004
JASA paper about "Getting it right."}
\usage{
as_theta(x)

as_theta(x)
}
\description{
the following two functions create the successive-conditional and
marginal-conditional sampling needed for comparison per John Geweke's 2004
JASA paper about "Getting it right."

the following two functions create the successive-conditional and
marginal-conditional sampling needed for comparison per John Geweke's 2004
JASA paper about "Getting it right."
}
