% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{default_glm_classifier}
\alias{default_glm_classifier}
\title{default_glm_classifier}
\usage{
default_glm_classifier()
}
\value{
returns a list of two functions: train and test.
}
\description{
Runs logistic regression on marginal conditional draws and successive conditional draws.
}
