% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{default_rforest_classifier}
\alias{default_rforest_classifier}
\title{default_rforest_classifier}
\usage{
default_rforest_classifier(ntree = 500)
}
\arguments{
\item{ntree}{Number of trees.}
}
\value{
returns a list of two functions: train and predict.
}
\description{
Runs a Random Forest classifier.
}
