% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{default_svm_classifier}
\alias{default_svm_classifier}
\title{default_svm_classifier}
\usage{
default_svm_classifier(kernel = "radial", cost = 1, gamma = NULL)
}
\arguments{
\item{kernel}{Kernel type.}

\item{cost}{Cost parameter.}

\item{gamma}{Gamma parameter (used for radial kernels).}
}
\value{
returns a list of two functions: train and predict.
}
\description{
Runs a Support Vector Machine classifier.
}
