% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prebaked-samplers.R
\name{doucet_build_D}
\alias{doucet_build_D}
\title{doucet_build_D}
\usage{
doucet_build_D(omega_k, n)
}
\value{
Returns the Fourier design matrix.
}
\description{
Constructs the Fourier design matrix used by the Andrieu and Doucet
reversible-jump MCMC example.
}
