% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prebaked-samplers.R
\name{doucet_inner_update}
\alias{doucet_inner_update}
\title{doucet_inner_update}
\usage{
doucet_inner_update(omega, a, sigma2, y, prior_hyper_params)
}
\value{
Returns the updated model parameters.
}
\description{
Performs the within-model Gibbs and Metropolis-Hastings updates for the
reversible-jump MCMC sampler.
}
