% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prebaked-samplers.R
\name{doucet_likelihood_sampler}
\alias{doucet_likelihood_sampler}
\title{doucet_likelihood_sampler}
\usage{
doucet_likelihood_sampler(n_obs, theta, prior_hyper_params)
}
\arguments{
\item{n_obs}{Number of observations to generate.}

\item{theta}{Model parameters.}

\item{prior_hyper_params}{List containing the prior hyperparameters.}
}
\value{
Returns a numeric vector of simulated observations.
}
\description{
Simulates observations from the Fourier regression model used in the
Andrieu and Doucet example.
}
