% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prebaked-samplers.R
\name{doucet_prior_sampler}
\alias{doucet_prior_sampler}
\title{doucet_prior_sampler}
\usage{
doucet_prior_sampler(prior_hyper_params)
}
\arguments{
\item{prior_hyper_params}{List containing the prior hyperparameters.}
}
\value{
Returns a named vector containing the sampled model parameters.
}
\description{
Draws parameters from the prior distribution for the Andrieu and Doucet
reversible-jump MCMC example.
}
