% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prebaked-samplers.R
\name{doucet_yPy}
\alias{doucet_yPy}
\title{doucet_yPy}
\usage{
doucet_yPy(omega_k, y, hyper)
}
\value{
Returns the computed quadratic form.
}
\description{
Computes the quadratic form used in the Andrieu and Doucet reversible-jump
MCMC algorithm.
}
