% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geweke.R
\name{geweke_sc_draws}
\alias{geweke_sc_draws}
\title{geweke_sc_draws}
\usage{
geweke_sc_draws(
  prior_sampler,
  likelihood_sampler,
  posterior_sampler,
  n_obs,
  n_draws = 10000,
  n_burn = 1000,
  n_thin = 5,
  prior_hyper_params
)
}
\arguments{
\item{prior_sampler}{The input for your prior sampler. Takes as input: prior_hyper_params. Returns: a named object (list or named vector) where names() gives parameter names and length() gives parameter count.}

\item{likelihood_sampler}{The input for your likelihood sampler. Takes as input: (n_obs, theta, prior_hyper_params). Returns: a plain numeric vector of length n_obs.}

\item{posterior_sampler}{The input for your posterior sampler. Takes as input: (n_draws, y, prior_hyper_params, init = NULL). First line of the function body should always be theta <- if (is.null(init)) <your_default> else unlist(init). Returns: a matrix with n_draws rows and named columns matching the names from prior_sampler.}

\item{n_obs}{Number of observations in each simulated dataset.}

\item{n_draws}{Number of draws from your joint. Default is 10,000 draws to ensure MCMC chain convergence.}

\item{n_burn}{Number of the first draws we discard so that the distribution can converge to an accurate sampler from our intended distribution.}

\item{n_thin}{We skip draws in the chain and only take every \code{n_thin}-th draw (e.g., every 5th draw) so that the dependency of the MCMC does not effect the distribution as much.}

\item{prior_hyper_params}{The list of prior hyperparameters required for your MCMC samplers to work as intended.}
}
\value{
Returns a list: list(draws = data.frame(draws_matrix),wide = data.frame(draws_matrix_wide), theta = data.frame(theta_matrix), y = data.frame(y_matrix)), which contains all draws from your simulated distribution.
}
\description{
The successive conditional draws from your joint.
}
