% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prebaked-samplers.R
\name{iidnorm_posterior_sampler}
\alias{iidnorm_posterior_sampler}
\title{iidnorm_posterior_sampler}
\usage{
iidnorm_posterior_sampler(
  n_draws,
  y,
  prior_hyper_params,
  init = c(mean(y), var(y))
)
}
\arguments{
\item{n_draws}{Number of posterior draws.}

\item{y}{Observed data.}

\item{prior_hyper_params}{List containing the prior hyperparameters.}

\item{init}{Initial values for the Markov chain.}
}
\value{
Returns a matrix of posterior draws.
}
\description{
Runs a Gibbs sampler for the IID Normal model with conjugate priors.
}
