% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prebaked-samplers.R
\name{iidnorm_prior_sampler}
\alias{iidnorm_prior_sampler}
\title{iidnorm_prior_sampler}
\usage{
iidnorm_prior_sampler(prior_hyper_params)
}
\arguments{
\item{prior_hyper_params}{List containing the prior hyperparameters.}
}
\value{
Returns a named list containing the sampled mean and variance.
}
\description{
Draws a sample from the conjugate Normal-Inverse-Gamma prior for the
IID Normal model.
}
