% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geweke.R
\name{keep_draw}
\alias{keep_draw}
\title{the following two functions create the successive-conditional and
marginal-conditional sampling needed for comparison per John Geweke's 2004
JASA paper about "Getting it right."}
\usage{
keep_draw(draw_num, n_burn, n_thin)
}
\description{
the following two functions create the successive-conditional and
marginal-conditional sampling needed for comparison per John Geweke's 2004
JASA paper about "Getting it right."
}
