% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{likelihood_sampler}
\alias{likelihood_sampler}
\title{likelihood_sampler}
\usage{
likelihood_sampler(n_obs, prior_samp, likelihood_input)
}
\arguments{
\item{n_obs}{number of observations}

\item{prior_samp}{the value drawn from the prior}

\item{likelihood_input}{function that tells the code what distribution to draw from and how many times}
}
\description{
Draws N times from the likelihood using the draw from the prior
NOTE: format the output later
}
