% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prebaked-samplers.R
\name{linreg_likelihood_sampler}
\alias{linreg_likelihood_sampler}
\title{linreg_likelihood_sampler}
\usage{
linreg_likelihood_sampler(n_obs, theta, prior_hyper_params)
}
\arguments{
\item{n_obs}{Number of observations to generate.}

\item{theta}{Named list containing the regression coefficients and variance.}

\item{prior_hyper_params}{List containing the design matrix and prior hyperparameters.}
}
\value{
Returns a numeric vector of simulated responses.
}
\description{
Simulates responses from a Bayesian linear regression model.
}
