% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prebaked-samplers.R
\name{linreg_posterior_sampler}
\alias{linreg_posterior_sampler}
\title{linreg_posterior_sampler}
\usage{
linreg_posterior_sampler(n_draws, y, prior_hyper_params, init = NULL)
}
\arguments{
\item{n_draws}{Number of posterior draws.}

\item{y}{Observed response vector.}

\item{prior_hyper_params}{List containing the prior hyperparameters.}

\item{init}{Initial values for the Markov chain.}
}
\value{
Returns a matrix of posterior draws.
}
\description{
Runs a Gibbs sampler for Bayesian linear regression with conjugate priors.
}
