% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prebaked-samplers.R
\name{linreg_prior_sampler}
\alias{linreg_prior_sampler}
\title{linreg_prior_sampler}
\usage{
linreg_prior_sampler(prior_hyper_params)
}
\arguments{
\item{prior_hyper_params}{List containing the prior hyperparameters.}
}
\value{
Returns a named list containing the sampled coefficients and variance.
}
\description{
Draws regression coefficients and the error variance from the Bayesian
linear regression prior.
}
