% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfunctions.R
\name{make_test_functions}
\alias{make_test_functions}
\title{make_test_functions}
\usage{
make_test_functions(...)
}
\arguments{
\item{...}{a list of all used test functions.}
}
\description{
creates a list of all test functions
}
