% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{posterior_sampler}
\alias{posterior_sampler}
\title{posterior_sampler}
\usage{
posterior_sampler(y, n_obs, posterior_input_function, prior_hyperparams)
}
\arguments{
\item{y}{the vector of likelihood draws}

\item{n_obs}{number of observations}

\item{posterior_input_function}{whatever the user thinks the posterior draws are supposed to be.}

\item{prior_hyperparams}{the vector of hyperparameters}
}
\description{
The posterior given to us to validate, as opposed to the one we approximate with gibbs.
NOTE: format the output later
}
