% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prebaked-samplers.R
\name{primiceri_posterior_sampler}
\alias{primiceri_posterior_sampler}
\title{primiceri_posterior_sampler}
\usage{
primiceri_posterior_sampler(
  n_draws,
  y,
  prior_hyper_params,
  init = NULL,
  buggy = FALSE
)
}
\arguments{
\item{n_draws}{Number of posterior draws.}

\item{y}{Observed data.}

\item{prior_hyper_params}{List containing the prior hyperparameters.}

\item{init}{Initial values for the Markov chain.}

\item{buggy}{Logical indicating whether to run the intentionally flawed sampler.}
}
\value{
Returns a matrix of posterior draws.
}
\description{
Runs the posterior sampler for the Primiceri (2005) stochastic volatility
example.
}
