% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prebaked-samplers.R
\name{primiceri_prior_sampler}
\alias{primiceri_prior_sampler}
\title{primiceri_prior_sampler}
\usage{
primiceri_prior_sampler(prior_hyper_params)
}
\arguments{
\item{prior_hyper_params}{List containing the prior hyperparameters.}
}
\value{
Returns the sampled model parameters.
}
\description{
Draws parameters from the prior distribution for the Primiceri (2005)
stochastic volatility example.
}
