% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prior_sampler}
\alias{prior_sampler}
\title{prior_sampler}
\usage{
prior_sampler(prior_input, prior_hyperparams)
}
\arguments{
\item{prior_input}{the function that details the actual distribution function}

\item{prior_hyperparams}{the vector of hyperparameters}
}
\description{
Draws 1 from the prior. relies on having no other inputs than the function and drawing from it once
}
