% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rank_kl_divergence}
\alias{rank_kl_divergence}
\title{Rank KL Divergence}
\usage{
rank_kl_divergence(results, L)
}
\arguments{
\item{results}{the res object}

\item{L}{number of posterior draws}
}
\description{
Test function for KL Divergence
}
