% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rank_ks_test}
\alias{rank_ks_test}
\title{Rank KS Test}
\usage{
rank_ks_test(results, L)
}
\arguments{
\item{results}{the res object}

\item{L}{number of posterior draws}
}
\description{
Test function for KS test
}
