% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rank_mean_test}
\alias{rank_mean_test}
\title{Rank Mean Test}
\usage{
rank_mean_test(results, L)
}
\arguments{
\item{results}{the res object}

\item{L}{number of posterior draws}
}
\description{
Test function for mean
}
