% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rank_variance_test}
\alias{rank_variance_test}
\title{Rank Variance Test}
\usage{
rank_variance_test(results, L)
}
\arguments{
\item{results}{the res object}

\item{L}{number of posterior draws}
}
\description{
Test function for varience
}
