% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc.R
\name{ranks_to_sbc_results}
\alias{ranks_to_sbc_results}
\title{ranks_to_sbc_results}
\usage{
ranks_to_sbc_results(ranked)
}
\arguments{
\item{ranked}{a bayescheckr_ranks object for conversion}
}
\value{
returns an SBC_results object for plotting
}
\description{
Convert a bayescheckr_ranks object to an SBC_results object for plotting
}
