% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfunctions.R
\name{recompute_ranks}
\alias{recompute_ranks}
\title{Test functions for SBC. We recompute ranks manually because the original
implementation in the SBC package was slow.}
\usage{
recompute_ranks(result, test_fns)
}
\arguments{
\item{result}{is created as \code{result <- run_sbc(...)}}
}
\description{
Test functions for SBC. We recompute ranks manually because the original
implementation in the SBC package was slow.
}
