% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{run_distributional_shift}
\alias{run_distributional_shift}
\title{run_distributional_shift}
\usage{
run_distributional_shift(
  direct_draws,
  gibbs_draws,
  classifier = NULL,
  train_frac = 0.8,
  n_perm = 20,
  alpha = 0.05,
  dist = "normal"
)
}
\arguments{
\item{direct_draws}{Direct draws from the joint distribution p(y, theta)}

\item{gibbs_draws}{Gibbs sampler draws from a simulated joint made up of successive conditional draws}

\item{classifier}{The algorithm used for classifying which proportion of the data is from which distribution. Should output a matrix of predictions.}

\item{train_frac}{The fraction of draws from your sampler that you want your classifier to train on. The remainder will be tested on for accuracy.}

\item{n_perm}{Used to determine feature importance, does not effect the classifier.}

\item{alpha}{Used to bound the P value of the C2ST test on our classifier.}
}
\value{
returns a list of the form: list(classifier_accuracy = accuracy,c2st = c2st, feature_importance = feature_importance).
}
\description{
Runs an input classifier to determine whether the two distributions can be determined as different by the machine or not.
Ideally you want 50\% accuracy with a uniform feature importance.
In the case of RJMCMC or other sampling methods that output samples in varying dimensions you may have trouble with some classification tools. The solution usualy is to put all samples in a max dimensional matrix and just leave the empty dimentions blank for low dimention samples.
In the case of high dimensional samples (I.E. 15+ dimensions) sometimes your classifier will be unable to work properly, such as SVM. In this case you will need a different classification method. We recommend xgboost.
}
