% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geweke.R
\name{run_geweke_prebaked}
\alias{run_geweke_prebaked}
\title{run_gewke_prebaked}
\usage{
run_geweke_prebaked(
  model,
  n_obs,
  n_draws = 10000,
  n_burn = 1000,
  n_thin = 5,
  prior_hyper_params = NULL,
  test_fns = NULL,
  parallelize = TRUE
)
}
\arguments{
\item{model}{select an option from: "iidnorm", "linreg", "doucet", "primiceri". All of which are our previously tested examples.}

\item{n_draws}{Number of posterior draws produced for each simulated dataset}

\item{n_burn}{Number of the first draws we discard so that the distribution can converge to an acurate sampler from our intended distribution}

\item{n_thin}{Number of draws we skip each time we draw so that the dependancy of the MCMC does not effect the distribution as much.}

\item{prior_hyper_params}{The list of prior hyperparameters reqired for your MCMC samplers to work as intended}

\item{test_fns}{A list of all the test functions that you want to run on your sampler}

\item{parallelize}{A Boolean that allows the user to choose if they want to use all their computer cores to speed up the computation.}
}
\value{
Returns a list: (list(direct = direct_draws, gibbs = gibbs_draws)). These are your gibbs and direct draws and what we are checking with the geweke thing.
}
\description{
Runs an example geweke from rebuilt and tested models
}
