% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc.R
\name{run_sbc}
\alias{run_sbc}
\title{The call to run SBC on your sampler.}
\usage{
run_sbc(
  prior_sampler,
  likelihood_sampler,
  posterior_sampler,
  n_sims,
  n_obs,
  n_draws = 10000,
  prior_hyper_params,
  test_fns,
  parallelize = FALSE,
  globals = NULL
)
}
\arguments{
\item{prior_sampler}{The input for your prior sampler}

\item{likelihood_sampler}{The input for your likelihood sampler}

\item{posterior_sampler}{The input for your posterior sampler.}

\item{n_sims}{Number of independent SBC simulations}

\item{n_obs}{Number of observations in each simulated dataset}

\item{n_draws}{Number of posterior draws produced for each simulated dataset}

\item{prior_hyper_params}{The list of prior hyperparameters reqired for your MCMC samplers to work as intended}

\item{test_fns}{A list of all the test functions that you want to run on your sampler}

\item{parallelize}{A Boolean that allows the user to choose if they want to use all their computer cores to speed up the computation.}

\item{globals}{Extra objects to export to parallel workers}
}
\value{
Returns a list: list(sbc_result = res, dataset = dataset, test_fns = test_fns), class = "bayescheckr_sbc") which then you can call upon using your plot functions.
}
\description{
This is done via an external package called SBC.
}
