% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc.R
\name{run_sbc_prebaked}
\alias{run_sbc_prebaked}
\title{run_sbc_prebaked}
\usage{
run_sbc_prebaked(
  model,
  n_sims,
  n_obs,
  n_draws = 10000,
  prior_hyper_params = NULL,
  test_fns = NULL,
  parallelize = FALSE,
  globals = NULL
)
}
\arguments{
\item{model}{select an option from: "iidnorm", "linreg", "doucet", "primiceri". All of which are our previously tested examples.}

\item{n_sims}{Number of independent SBC simulations}

\item{n_obs}{Number of observations in each simulated dataset}

\item{n_draws}{Number of posterior draws produced for each simulated dataset}

\item{prior_hyper_params}{The list of prior hyperparameters reqired for your MCMC samplers to work as intended}

\item{test_fns}{A list of all the test functions that you want to run on your sampler}

\item{parallelize}{A Boolean that allows the user to choose if they want to use all their computer cores to speed up the computation.}

\item{globals}{Extra objects to export to parallel workers}
}
\value{
Returns a list: list(sbc_result = res, dataset = dataset, test_fns = test_fns), class = "bayescheckr_sbc") which then you can call upon using your plot functions.
}
\description{
Runs an example SBC from rebuilt and tested models
}
