% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geweke.R
\name{tabulate_geweke_tests}
\alias{tabulate_geweke_tests}
\title{tabulate_geweke_tests}
\usage{
tabulate_geweke_tests(draws, test_functions, tests_applied = FALSE)
}
\arguments{
\item{draws}{Your matrix of draws.}

\item{test_functions}{Your test functions}

\item{tests_applies}{If you have already applied test functions then skip the step where you apply them.}
}
\value{
Returns a results matrix for all your test functions
}
\description{
Takes your draws and test functions and applies them if needed, then takes the results and outputs them as a matrix.
}
