% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geweke.R
\name{as_theta}
\alias{as_theta}
\title{the following two functions create the successive-conditional and
marginal-conditional sampling needed for comparison per John Geweke's 2004
JASA paper about "Getting it right."
as_theta}
\usage{
as_theta(x)
}
\arguments{
\item{x}{input theta.}
}
\value{
output theta in the correct format.
}
\description{
Makes sure theta is constructed in a way that makes it robust to different access methods.
}
