% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{default_xgb_classifier}
\alias{default_xgb_classifier}
\title{default_xgb_classifier}
\usage{
default_xgb_classifier(nrounds = 100)
}
\arguments{
\item{nrounds}{How many "cuts" XGBoost.}
}
\value{
returns a list of two functions: train and test.
}
\description{
Runs XG boost classifier instead of the default logistic regression.
}
