% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geweke.R
\name{geweke_mc_draws}
\alias{geweke_mc_draws}
\title{geweke_mc_draws}
\usage{
geweke_mc_draws(
  prior_sampler,
  likelihood_sampler,
  n_obs,
  n_draws = 10000,
  prior_hyper_params,
  parallelize = TRUE
)
}
\arguments{
\item{prior_sampler}{The input for your prior sampler}

\item{likelihood_sampler}{The input for your likelihood sampler}

\item{n_obs}{Number of observations in each simulated dataset}

\item{n_draws}{Number of draws from your joint}

\item{prior_hyper_params}{The list of prior hyperparameters required for your MCMC samplers to work as intended}

\item{posterior_sampler}{The input for your posterior sampler.}
}
\value{
Returns a list: list(draws = data.frame(draws_matrix),wide = data.frame(draws_matrix_wide), theta = data.frame(theta_matrix), y = data.frame(y_matrix)), which contains all draws from your simulated distribution
}
\description{
The marginal conditional draws from your joint.
}
