% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geweke.R
\name{geweke_sc_draws}
\alias{geweke_sc_draws}
\title{the following two functions create the successive-conditional and
marginal-conditional sampling needed for comparison per John Geweke's 2004
JASA paper about "Getting it right."}
\usage{
geweke_sc_draws(
  prior_sampler,
  likelihood_sampler,
  posterior_sampler,
  prior_hyper_params,
  n_draws,
  n_obs
)
}
\description{
the following two functions create the successive-conditional and
marginal-conditional sampling needed for comparison per John Geweke's 2004
JASA paper about "Getting it right."
}
