% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geweke.R
\name{geweke_sc_draws}
\alias{geweke_sc_draws}
\title{geweke_sc_draws}
\usage{
geweke_sc_draws(
  prior_sampler,
  likelihood_sampler,
  posterior_sampler,
  n_obs,
  n_draws = 10000,
  n_burn = 1000,
  n_thin = 5,
  prior_hyper_params
)
}
\arguments{
\item{prior_sampler}{The input for your prior sampler}

\item{likelihood_sampler}{The input for your likelihood sampler}

\item{posterior_sampler}{The input for your posterior sampler.}

\item{n_obs}{Number of observations in each simulated dataset}

\item{n_burn}{Number of the first draws we discard so that the distribution can converge to an accurate sampler from our intended distribution}

\item{prior_hyper_params}{The list of prior hyperparameters required for your MCMC samplers to work as intended}

\item{thin}{Number of draws we skip each time we draw so that the dependency of the MCMC does not effect the distribution as much.}
}
\value{
Returns a list: list(draws = data.frame(draws_matrix),wide = data.frame(draws_matrix_wide), theta = data.frame(theta_matrix), y = data.frame(y_matrix)), which contains all draws from your simulated distribution
}
\description{
The successive conditional draws from your joint.
}
