% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geweke_list.R, R/geweke_matrix.R
\name{geweke_test}
\alias{geweke_test}
\title{Now moving to test functions: difference in means testing (per Geweke 2004)
and KS testing, for z-score comparison, among other tests}
\usage{
geweke_test(g_mc, g_sc)

geweke_test(g_mc, g_sc)
}
\description{
Now moving to test functions: difference in means testing (per Geweke 2004)
and KS testing, for z-score comparison, among other tests

Now moving to test functions: difference in means testing (per Geweke 2004)
and KS testing, for z-score comparison, among other tests
}
