% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfunctions.R
\name{make_test_functions}
\alias{make_test_functions}
\title{We require that the user gives test functions inside a NAMED list.}
\usage{
make_test_functions(...)
}
\description{
\if{html}{\out{<div class="sourceCode r">}}\preformatted{test_fns <- make_test_functions(
  mu_sq = function(theta, y) theta["mu"]^2,
  product = function(theta, y) theta["mu"] * theta["sigmasq"],
  log_like = function(theta, y)
    sum(dnorm(y, theta["mu"],
              sqrt(theta["sigmasq"]),
              log = TRUE))
)
}\if{html}{\out{</div>}}
}
