% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfunctions.R
\name{make_test_functions}
\alias{make_test_functions}
\title{We require that the user gives test functions inside a NAMED list.
Something like:
test_fns <- list(
mu_sq = function(theta, y) theta\link{"mu"}^2,
product = function(theta, y) theta\link{"mu"} * theta\link{"sigmasq"},
log_like = function(theta, y) sum(dnorm(y, theta\link{"mu"},
sqrt(theta\link{"sigmasq"}), log = TRUE))
)}
\usage{
make_test_functions(...)
}
\description{
We require that the user gives test functions inside a NAMED list.
Something like:
test_fns <- list(
mu_sq = function(theta, y) theta\link{"mu"}^2,
product = function(theta, y) theta\link{"mu"} * theta\link{"sigmasq"},
log_like = function(theta, y) sum(dnorm(y, theta\link{"mu"},
sqrt(theta\link{"sigmasq"}), log = TRUE))
)
}
