% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc.R
\name{ranks_to_sbc_results}
\alias{ranks_to_sbc_results}
\title{A full picture of what run_sbc() returns:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{result <- run_sbc(...)
}\if{html}{\out{</div>}}}
\usage{
ranks_to_sbc_results(ranked)
}
\description{
result$sbc_result                    # native SBC_results, use with SBC:: functions
result$dataset$variables             # n_sims x n_params matrix of theta_tilde draws
result$dataset$generated[[i]]$y      # y vector for simulation i
result$sbc_result$fits[[i]]          # n_draws x n_params posterior draws for simulation i

\if{html}{\out{<div class="sourceCode r">}}\preformatted{
[i]: R:i
[i]: R:i
}\if{html}{\out{</div>}}
}
