% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geweke.R
\name{run_geweke}
\alias{run_geweke}
\title{run_gewke}
\usage{
run_geweke(
  prior_sampler,
  likelihood_sampler,
  posterior_sampler,
  n_obs,
  n_draws = 10000,
  n_burn = 1000,
  n_thin = 5,
  prior_hyper_params,
  test_fns = NULL,
  parallelize = TRUE
)
}
\arguments{
\item{prior_sampler}{The input for your prior sampler}

\item{likelihood_sampler}{The input for your likelihood sampler}

\item{posterior_sampler}{The input for your posterior sampler.}

\item{n_obs}{Number of observations in each simulated dataset}

\item{n_burn}{Number of the first draws we discard so that the distribution can converge to an acurate sampler from our intended distribution}

\item{prior_hyper_params}{The list of prior hyperparameters reqired for your MCMC samplers to work as intended}

\item{test_fns}{A list of all the test functions that you want to run on your sampler}

\item{parallelize}{A Boolean that allows the user to choose if they want to use all their computer cores to speed up the computation.}

\item{thin}{Number of draws we skip each time we draw so that the dependancy of the MCMC does not effect the distribution as much.}
}
\value{
Returns a list: (list(direct = direct_draws, gibbs = gibbs_draws)). These are your gibbs and direct draws and what we are checking with the geweke thing.
}
\description{
Puts everything together and creates the SC and MC draws for the test to use.
}
